feat: add eslint-config-airbnb-base
