+++ .eslintrc
@@ -1,0 +1,8 @@
+{
+  "extends": [
+    "airbnb-base"
+  ],
+  "env": {
+    "browser": true
+  }
+}
