+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "virtual-keyboard",
+  "version": "1.0.0",
+  "description": "",
+  "main": "src/object.js",
+  "scripts": {
+    "lint": "eslint ./src",
+    "lint:fix": "eslint ./src --fix"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Kirsawka/virtual-keyboard.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/Kirsawka/virtual-keyboard/issues"
+  },
+  "homepage": "https://github.com/Kirsawka/virtual-keyboard#readme",
+  "devDependencies": {
+    "eslint-config-airbnb": "^19.0.4"
+  }
+}
